+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 6.0.0
-* Updated Native `Android` SDK:
+* Updated Native `Android` and `iOS` SDKs:
     * Android Native SDK Update 5.2.0 - [Android Version History](https://github.com/BranchMetrics/android-branch-deep-linking-attribution/releases)
     * iOS Native SDK Update 1.43.0 - [iOS Version History](https://github.com/BranchMetrics/ios-branch-deep-linking-attribution/releases)